Drill guy entering/exiting states
Added basic walking state (just waits), and then a enter state that spawns a hole and an exit state that also spawns a hole. Also, added a drill path class that is basically a bezier curve to make underground movement interesting, an created test scene for the digging push_back and state behavior loop.